requirements.txt
asgiref==3.3.4
Django==3.2.3
django-environ==0.4.5
django-rest-framework==0.1.0
djangorestframework==3.12.4
environ==1.0
psycopg2==2.8.6
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
uritemplate==3.0.1
